Fix random number generation API compatibility issues
- Remove duplicate Rng imports
- Use rand_core::OsRng.fill_bytes instead of getrandom crate
- Use RngCore::fill_bytes for ChaCha20Rng
- Fix rand::random::<usize>() to use u32 cast to usize
- Make threshold variable mutable in weighted selection
- Add comprehensive fix summary documentation

+++ src/core/server/smart_routing.rs
@@ -7,7 +7,6 @@ use std::sync::Arc;
 use tokio::sync::{Mutex, RwLock};
 use serde::{Serialize, Deserialize};
 use rand::Rng;
-use rand::Rng;
 use crate::error::{VantisError, Result};
 
 /// Metric used to evaluate network paths for routing decisions
@@ -375,7 +374,7 @@ impl SmartRoutingManager {
             let exploration_value: f64 = rand::random();
             if exploration_value < self.config.exploration_rate {
                 // Explore: choose random path
-                let random_index = rand::random::<usize>() % paths.len();
+                let random_index = (rand::random::<u32>() as usize) % paths.len();
                 let random_path = paths
                     .get(random_index)
                     .ok_or_else(|| VantisError::InvalidPeer("No paths available".to_string()))?;
